pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel", "tomli>=2.0.1; python_version < '3.11'"]
build-backend = "setuptools.build_meta"

[project]
name = "autoencoding-pcs"
version = "1.0.0"
description = "Tractable Representation Learning with Probabilistic Circuits"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE.md" }
authors = [{ name = "Steven Braun" }]
dependencies = [
    "setuptools==68.2.2",
    "hydra-core==1.3.2",
    "icecream==2.1.3",
    "matplotlib==3.8.1",
    "seaborn==0.13.0",
    "numpy==1.26.1",
    "omegaconf==2.3.0",
    "Pillow==10.1.0",
    "lightning==2.4.0",
    "rich==13.6.0",
    "rtpt==0.0.4",
    "scikit_learn==1.3.1",
    "timm==0.9.7",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "torchaudio==2.5.1",
    "tqdm==4.66.1",
    "wandb==0.18.5",
    "hydra_colorlog==1.2.0",
    "fast_pytorch_kmeans==0.2.0.1",
    "tensorboard==2.14.1",
    "parameterized==0.9.0",
    "torchsummary==1.5.1",
    "torchinfo==1.8.0",
    "imageio==2.34.1",
    "zuko==1.2.0",
    "pytorch-msssim==1.0.0",
    "lmdb==1.5.1",
    "umap-learn==0.5.7",
    "pytorch_warmup==0.2.0",
    "mnist1d==0.0.2.post1",
    "openpyxl==3.1.5",
    "xlrd==2.0.1",
    "h5py==3.13.0",
    "runitall @ git+https://github.com/braun-steven/runitall.git",
]

[tool.setuptools.packages.find]
include = ["simple_einet", "simple_einet.*"]

[tool.black]
line-length = 120
